+++ waterfall-cad/src/Waterfall/Path.hs
@@ -83,7 +83,7 @@ pathFromTo3D :: [V3 Double -> (V3 Double, Path)] -> V3 Double -> (V3 Double, Pat
 pathFromTo3D = pathFromTo
 
 -- | `pathEndpoints`, with the type fixed to `Path` 
-pathEndpoints3D :: Path -> (V3 Double, V3 Double)
+pathEndpoints3D :: Path -> Maybe (V3 Double, V3 Double)
 pathEndpoints3D = pathEndpoints
 
 -- | `closeLoop` with the type fixed to `Path`