+++ waterfall-cad/src/Waterfall/Sweep.hs
@@ -4,6 +4,7 @@ module Waterfall.Sweep
 
 import Waterfall.Internal.Solid (Solid (..), acquireSolid, solidFromAcquire)
 import Waterfall.Internal.Path (Path (..))
+import Waterfall.Internal.Path.Common (RawPath (..))
 import Waterfall.Internal.Edges (wireTangent, wireEndpoints)
 import Waterfall.Internal.Finalizers (toAcquire)
 import Waterfall.Transforms (rotate, translate)
@@ -16,6 +17,7 @@ import Control.Monad.IO.Class (liftIO)
 import Foreign.Ptr
 import Linear (V3, normalize, unit, _x, _z, nearZero, cross, dot)
 import Data.Acquire (Acquire)
+import qualified Waterfall.Solids as Solids
 
 rotateFace :: V3 Double -> Ptr TopoDS.Shape -> Acquire (Ptr TopoDS.Shape)
 rotateFace v face = 
@@ -33,7 +35,7 @@ positionFace p = acquireSolid . translate p . solidFromAcquire . pure
 
 -- | Sweep a 2D `Shape` along a `Path`, constructing a `Solid`
 sweep :: Path -> Shape -> Solid
-sweep (Path theRawPath) (Shape theRawShape) = solidFromAcquire $ do
+sweep (Path (ComplexRawPath theRawPath)) (Shape theRawShape) = solidFromAcquire $ do
     path <- toAcquire theRawPath
     shape <- toAcquire theRawShape
     tangent <- liftIO $ wireTangent path
@@ -41,3 +43,4 @@ sweep (Path theRawPath) (Shape theRawShape) = solidFromAcquire $ do
     adjustedFace <- positionFace start =<< rotateFace tangent shape
     builder <- MakePipe.fromWireAndShape path adjustedFace
     MakeShape.shape (upcast builder)
+sweep _ _ = Solids.nowhere