+++ waterfall-cad/src/Waterfall/TwoD/Shape.hs
@@ -16,12 +16,15 @@ import qualified OpenCascade.BRepBuilderAPI.MakeFace as MakeFace
 import OpenCascade.Inheritance (upcast)
 import Linear (unit, _x, _y, zero, V2 (..))
 import Waterfall.Path.Common (pathFrom, arcViaTo, lineTo)
+import Waterfall.Internal.Path.Common (RawPath(ComplexRawPath))
 
 -- | Construct a 2D Shape from a closed path 
 fromPath :: Path2D -> Shape
-fromPath (Path2D r) = Shape . unsafeFromAcquire  $ do
+fromPath (Path2D (ComplexRawPath r)) = Shape . unsafeFromAcquire  $ do
     p <- toAcquire r
     upcast <$> (MakeFace.face =<< MakeFace.fromWire p False)
+fromPath _ = Shape . unsafeFromAcquire $
+    upcast <$> (MakeFace.face =<< MakeFace.new)
 
 -- | Get the paths back from a 2D shape
 -- 
@@ -31,7 +34,7 @@ fromPath (Path2D r) = Shape . unsafeFromAcquire  $ do
 -- shapePaths . fromPath ≡ pure
 -- @
 shapePaths :: Shape -> [Path2D] 
-shapePaths (Shape r) = fmap Path2D . unsafeFromAcquire $ do
+shapePaths (Shape r) = fmap (Path2D . ComplexRawPath) . unsafeFromAcquire $ do
     s <- toAcquire r 
     allWires s