+++ waterfall-cad/CHANGELOG.md
@@ -16,9 +16,15 @@ and this project adheres to the
 - Add `matTransform` and `matTransform2D` methods to the `Transformable` and `Transformable2D` typeclasses, respectively
 - Add `shapePaths` to `Waterfall.TwoD.Shape`
 
+### Changed
+
+- Most functions in `Waterfall.Path.Common` now have an `Epsilon` constraint
+- `pathEndpoints`/`pathEndpoints3D`/`pathEndpoints2D` now returns a `Maybe` (in case of an empty path)
+
 ### Fixed
 
 - fixed behaviour when scaling with a unit vector (no scaling)
+- fixed `Path`/`Path2D` representation, so that the `Monoid` instance `mempty` value no longer generates crashes
 
 ## 0.4.0.0